requirements.txt
asgiref==3.8.1
Django==5.1.3
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
psycopg2-binary==2.9.10
sqlparse==0.5.2
tzdata==2024.2
